refactor: 기존 상속 구조로 재변경
