+++ src/main/java/chess/piece/King.java
@@ -1,15 +1,14 @@
-package chess.piece.directional.normal;
+package chess.piece;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import chess.board.Position;
-import chess.piece.Direction;
-import chess.piece.Piece;
-import chess.piece.Side;
 
-public class King extends NormalPiece {
+public class King extends Piece {
+
+    private final List<Direction> directions;
 
     private static final int KING_MOVE_COUNT = 1;
 
@@ -18,8 +17,7 @@ public class King extends NormalPiece {
         this.directions = initDirections();
     }
 
-    @Override
-    protected List<Direction> initDirections() {
+    private List<Direction> initDirections() {
         return new ArrayList<>(Arrays.asList(Direction.values()));
     }
 
@@ -34,5 +32,23 @@ public class King extends NormalPiece {
     @Override
     public Piece move(Position positionToMove) {
         return new King(positionToMove, this.side);
+    }
+
+    @Override
+    public boolean isNeedToCheckWhenDiagonalMove() {
+        return false;
+    }
+
+    @Override
+    public List<Position> getPaths(Position targetPosition) {
+        List<Position> paths = new ArrayList<>();
+        final Direction direction = position.getDirectionTo(targetPosition);
+        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
+        Position nextPosition = this.position;
+        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
+            nextPosition = nextPosition.getNextPosition(direction);
+            paths.add(nextPosition);
+        }
+        return paths;
     }
 }
