+++ src/test/java/chess/piece/KingTest.java
@@ -8,7 +8,6 @@ import chess.board.Rank;
 import java.util.Collections;
 import java.util.List;
 
-import chess.piece.directional.normal.King;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
