+++ src/test/java/chess/piece/PawnTest.java
@@ -8,8 +8,6 @@ import chess.board.Rank;
 import java.util.Collections;
 import java.util.List;
 
-import chess.piece.directional.normal.King;
-import chess.piece.directional.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
