+++ src/test/java/chess/piece/PieceTest.java
@@ -7,11 +7,6 @@ import java.util.stream.Stream;
 import chess.board.File;
 import chess.board.Position;
 import chess.board.Rank;
-import chess.piece.directional.*;
-import chess.piece.directional.normal.King;
-import chess.piece.directional.normal.longrange.Bishop;
-import chess.piece.directional.normal.longrange.Queen;
-import chess.piece.directional.normal.longrange.Rook;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
