+++ src/test/java/chess/piece/RookTest.java
@@ -7,7 +7,6 @@ import chess.board.Position;
 import chess.board.Rank;
 import java.util.List;
 
-import chess.piece.directional.normal.longrange.Rook;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
